+++ lib/linkedin_login.dart
@@ -5,3 +5,4 @@ export 'package:linkedin_login/src/linked_in_user_model.dart';
 
 export 'package:linkedin_login/src/linked_in_buttons.dart';
 export 'package:linkedin_login/src/linked_in_profile.dart';
+export 'package:linkedin_login/src/linked_in_auth_code.dart';