+++ lib/src/linked_in_authorization_webview.dart
@@ -5,6 +5,8 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter_webview_plugin/flutter_webview_plugin.dart';
 import 'package:http/http.dart';
+import 'package:linkedin_login/helper/global_variables.dart';
+import 'package:linkedin_login/helper/helper_methods.dart';
 import 'package:linkedin_login/src/linked_in_auth_response_wrapper.dart';
 import 'package:uuid/uuid.dart';
 
@@ -27,10 +29,6 @@ class LinkedInAuthorization extends StatefulWidget {
 }
 
 class _LinkedInAuthorizationState extends State<LinkedInAuthorization> {
-  final urlLinkedInAccessToken =
-      'https://www.linkedin.com/oauth/v2/accessToken';
-  final urlLinkedInAuthorization =
-      'https://www.linkedin.com/oauth/v2/authorization';
   final flutterWebViewPlugin = FlutterWebviewPlugin();
 
   StreamSubscription<String> _onUrlChanged;
@@ -53,7 +51,7 @@ class _LinkedInAuthorizationState extends State<LinkedInAuthorization> {
 
     flutterWebViewPlugin.close();
 
-    loginUrl = '$urlLinkedInAuthorization?'
+    loginUrl = '${GlobalVariables.URL_LINKED_IN_GET_AUTH_TOKEN}?'
         'response_type=code'
         '&client_id=${widget.clientId}'
         '&state=$clientState'
@@ -64,11 +62,10 @@ class _LinkedInAuthorizationState extends State<LinkedInAuthorization> {
     _onUrlChanged = flutterWebViewPlugin.onUrlChanged.listen((String url) {
       if (mounted && url.startsWith(widget.redirectUrl)) {
         flutterWebViewPlugin.stopLoading();
-        _getAuthorizationCode(url: url).then(
-          (AuthorizationCodeResponse result) {
-            widget.onCallBack(result);
-          },
-        );
+
+        AuthorizationCodeResponse authCode =
+            getAuthorizationCode(redirectUrl: url, clientState: clientState);
+        _getAccessToken(authorizationCode: authCode).then(widget.onCallBack);
       }
     });
   }
@@ -76,71 +73,34 @@ class _LinkedInAuthorizationState extends State<LinkedInAuthorization> {
   /// Method that will retrieve authorization code
   /// After auth code is received you can call API service to get access token
   /// from linkedIn
-  Future<AuthorizationCodeResponse> _getAuthorizationCode({String url}) async {
-    AuthorizationCodeResponse response;
-
-    // logic for parsing URL from your redirect url so that we can get access
-    // code
-
-    final List<String> parseUrl = url.split('?');
-
-    if (parseUrl.isNotEmpty) {
-      final List<String> queryPart = parseUrl.last.split('&');
-
-      if (queryPart.isNotEmpty && queryPart.first.startsWith('code')) {
-        final List<String> codePart = queryPart.first.split('=');
-        final List<String> statePart = queryPart.last.split('=');
-
-        if (statePart[1] == clientState) {
-          response = AuthorizationCodeResponse(
-            code: codePart[1],
-            state: statePart[1],
-          );
-        } else {
-          AuthorizationCodeResponse(
-            error: LinkedInErrorObject(
-              statusCode: HttpStatus.unauthorized,
-              description: 'State code is not valid: ${statePart[1]}',
-            ),
-            state: statePart[1],
-          );
-        }
-      } else if (queryPart.isNotEmpty && queryPart.first.startsWith('error')) {
-        response = AuthorizationCodeResponse(
-          error: LinkedInErrorObject(
-            statusCode: HttpStatus.unauthorized,
-            description: queryPart[1].split('=')[1].replaceAll('+', ' '),
-          ),
-          state: queryPart[2].split('=')[1],
-        );
-      }
-    }
-
+  Future<AuthorizationCodeResponse> _getAccessToken(
+      {AuthorizationCodeResponse authorizationCode}) async {
     // get access token based on code
-    if (response.code != null && response.code.isNotEmpty) {
-      final LinkedInTokenObject tokenObject = await _getAccessToken(response);
+    if (authorizationCode.code != null && authorizationCode.code.isNotEmpty) {
+      final LinkedInTokenObject tokenObject =
+          await _getUserProfile(authorizationCode);
 
       if (tokenObject.error == null || tokenObject.error.description.isEmpty) {
-        response.accessToken = tokenObject;
+        authorizationCode.accessToken = tokenObject;
       } else {
-        response.error = LinkedInErrorObject(
+        authorizationCode.error = LinkedInErrorObject(
           statusCode: HttpStatus.unauthorized,
           description: 'Failed to get user token',
         );
-        response.accessToken = null;
-        response.code = null;
-        response.state = null;
+        authorizationCode.accessToken = null;
+        authorizationCode.code = null;
+        authorizationCode.state = null;
       }
     }
 
     flutterWebViewPlugin.close();
 
-    return response;
+    return authorizationCode;
   }
 
   /// Method for getting token object for current user
   /// This method will return you token & expiration time for that token
-  Future<LinkedInTokenObject> _getAccessToken(
+  Future<LinkedInTokenObject> _getUserProfile(
       AuthorizationCodeResponse codeDetails) async {
     final Map<String, dynamic> body = {
       'grant_type': 'authorization_code',
@@ -150,7 +110,7 @@ class _LinkedInAuthorizationState extends State<LinkedInAuthorization> {
       'client_secret': widget.clientSecret,
     };
 
-    final response = await post(urlLinkedInAccessToken,
+    final response = await post(GlobalVariables.URL_LINKED_IN_GET_ACCESS_TOKEN,
         body: body,
         headers: {
           HttpHeaders.acceptHeader: 'application/json',